fix(database): handle potential errors and edge cases when closing database connection
- Add error handling for SQLITE_MISUSE errors
- Check if database is already closed or in a bad state
- Log appropriate error messages
- Ensure db is set to null in all cases
