+++ src/renderer/EverEtchApp.ts
@@ -1,10 +1,16 @@
-import { WordDocument, WordListItem, PaginationState, AssociatedWordsState } from './types.js';
+import { WordDocument, WordListItem } from './types.js';
 import { ProfileService } from './services/ProfileService.js';
 import { WordService } from './services/WordService.js';
 import { WordImportService, ImportProgress, ImportCallbacks } from './services/WordImportService.js';
 import { ToastManager } from './components/ToastManager.js';
 import { WordRenderer } from './components/WordRenderer.js';
 import { UIUtils } from './utils/UIUtils.js';
+import { WordManager } from './services/WordManager.js';
+import { AssociatedWordsManager } from './services/AssociatedWordsManager.js';
+import { ModalManager } from './components/ModalManager.js';
+import { GoogleDriveManager } from './services/GoogleDriveManager.js';
+import { ProtocolManager } from './services/ProtocolManager.js';
+import { EventManager } from './services/EventManager.js';
 import { generateGenerationId } from './utils/Common.js';
 
 // Constants for pagination
@@ -12,6 +18,7 @@ const WORDS_PAGE_SIZE = 10; // For release builds, set to 10, debug 5
 const ASSOCIATED_WORDS_PAGE_SIZE = 10; // For release builds, set to 10, debug 5
 
 export class EverEtchApp {
+  // Core services (keep existing)
   private profileService: ProfileService;
   private wordService: WordService;
   private wordImportService: WordImportService;
@@ -19,48 +26,65 @@ export class EverEtchApp {
   private wordRenderer: WordRenderer;
   private uiUtils: UIUtils;
 
-  // App state
-  private currentWord: WordDocument | null = null;
-  private currentGenerationId: string = '';
-  private streamingContent: string = '';
-  private words: WordListItem[] = [];
-  private wordsPagination: PaginationState = {
-    offset: 0,
-    pageSize: WORDS_PAGE_SIZE,
-    isLoading: false,
-    hasMore: true,
-    total: 0
-  };
-  private associatedWordsState: AssociatedWordsState = {
-    words: [],
-    offset: 0,
-    pageSize: ASSOCIATED_WORDS_PAGE_SIZE,
-    isLoading: false,
-    hasMore: true,
-    total: 0,
-    currentTag: '',
-    scrollObserver: null
-  };
-  private isSearchMode: boolean = false;
-  private isGenerating: boolean = false;
-  private isImporting: boolean = false;
-  private scrollObserver: IntersectionObserver | null = null;
-  private sortOrder: 'asc' | 'desc' = 'desc';
-  private selectedGoogleDriveFile: any = null;
+  // New extracted managers
+  private wordManager!: WordManager;
+  private associatedWordsManager!: AssociatedWordsManager;
+  private modalManager!: ModalManager;
+  private googleDriveManager!: GoogleDriveManager;
+  private protocolManager!: ProtocolManager;
+  private eventManager!: EventManager;
+
+  // Minimal app-level state (managers handle their own state)
 
   constructor() {
+    // Initialize core services
     this.profileService = new ProfileService();
     this.wordService = new WordService();
     this.toastManager = new ToastManager();
     this.uiUtils = new UIUtils();
     this.wordImportService = new WordImportService(this.wordService, this.toastManager);
-
     this.wordRenderer = new WordRenderer(this.wordService, this.toastManager);
 
-    // Set up event callbacks
-    this.setupWordRendererCallbacks();
+    // Initialize managers
+    this.initializeManagers();
 
+    // Initialize app
     this.initializeApp();
+  }
+
+  private initializeManagers(): void {
+    // Create associated words manager first (needed by word manager)
+    this.associatedWordsManager = new AssociatedWordsManager(
+      this.wordService,
+      this.toastManager,
+      this.wordRenderer,
+      this.uiUtils
+    );
+
+    // Create word manager
+    this.wordManager = new WordManager(
+      this.wordService,
+      this.toastManager,
+      this.wordRenderer,
+      this.uiUtils,
+      this.associatedWordsManager
+    );
+
+    // Create other managers
+    this.modalManager = new ModalManager(this.toastManager, this.profileService);
+    this.googleDriveManager = new GoogleDriveManager(this.toastManager, this.profileService);
+    this.protocolManager = new ProtocolManager(this.toastManager);
+
+    // Create event manager (depends on all other managers)
+    this.eventManager = new EventManager(
+      this.wordManager,
+      this.modalManager,
+      this.googleDriveManager,
+      this.profileService,
+      this.uiUtils,
+      this.wordImportService,
+      this.toastManager
+    );
   }
 
   private showLoadingOverlay(): void {
@@ -74,14 +98,15 @@ export class EverEtchApp {
   }
 
   private setupWordRendererCallbacks(): void {
-    this.wordRenderer.onTagClick = (tag: string) => this.loadAssociatedWords(tag);
-    this.wordRenderer.onSynonymClick = (synonym: string) => this.loadAssociatedWords(synonym);
-    this.wordRenderer.onAntonymClick = (antonym: string) => this.loadAssociatedWords(antonym);
-    this.wordRenderer.onAddWord = (word: WordDocument) => this.handleAddWord(word);
-    this.wordRenderer.onRefreshWord = (word: WordDocument) => this.handleRefreshWord(word);
-    this.wordRenderer.onDeleteWord = (word: WordDocument) => this.handleDeleteWord(word);
-    this.wordRenderer.onWriteRemark = (word: WordDocument) => this.handleWriteRemark(word);
-    this.wordRenderer.onWordSelect = (word: WordDocument | WordListItem) => this.selectWord(word);
+    // Delegate to word manager
+    this.wordRenderer.onTagClick = (tag: string) => this.associatedWordsManager.loadAssociatedWords(tag);
+    this.wordRenderer.onSynonymClick = (synonym: string) => this.associatedWordsManager.loadAssociatedWords(synonym);
+    this.wordRenderer.onAntonymClick = (antonym: string) => this.associatedWordsManager.loadAssociatedWords(antonym);
+    this.wordRenderer.onAddWord = (word: WordDocument) => this.wordManager.handleAddWord(word);
+    this.wordRenderer.onRefreshWord = (word: WordDocument) => this.wordManager.handleRefreshWord(word);
+    this.wordRenderer.onDeleteWord = (word: WordDocument) => this.wordManager.handleDeleteWord(word);
+    this.wordRenderer.onWriteRemark = (word: WordDocument) => this.wordManager.handleWriteRemark(word);
+    this.wordRenderer.onWordSelect = (word: WordDocument | WordListItem) => this.wordManager.selectWord(word);
   }
 
   private async initializeApp(): Promise<void> {
@@ -92,7 +117,7 @@ export class EverEtchApp {
       console.log('Loaded profiles:', this.profileService.getProfiles());
 
       // Load saved sort order BEFORE loading words
-      await this.loadSortOrder();
+      await this.wordManager.loadSortOrder();
 
       // Check if there's a last opened profile and switch to it with loading overlay
       const currentProfile = this.profileService.getCurrentProfile();
@@ -132,14 +157,16 @@ export class EverEtchApp {
         await this.loadWords();
       }      
 
+      // Set up word renderer callbacks
+      this.setupWordRendererCallbacks();
+
       // Set up event listeners
       this.setupEventListeners();
 
       // Load saved panel widths
       await this.uiUtils.loadPanelWidths();
 
-      // Update sort button icon after loading
-      this.updateSortButtonIcon();
+
 
       // Signal to main process that app is fully ready
       console.log('🎯 Renderer: App fully initialized, sending app-ready signal');
@@ -152,2437 +179,107 @@ export class EverEtchApp {
   }
 
   private async loadWords(): Promise<void> {
-    try {
-      // Reset pagination state
-      this.words = [];
-      this.wordsPagination = {
-        offset: 0,
-        pageSize: WORDS_PAGE_SIZE,
-        isLoading: false,
-        hasMore: true,
-        total: 0
-      };
-
-      // Clean up existing observer
-      if (this.scrollObserver) {
-        this.scrollObserver.disconnect();
-        this.scrollObserver = null;
-      }
-
-      // Clear the word list UI before loading new data
-      const wordList = document.getElementById('word-list')!;
-      if (wordList) {
-        wordList.innerHTML = '';
-      }
-
-      // Load first page first, then setup observer after content is rendered
-      await this.loadMoreWords();
-
-      // Setup scroll observer after content is rendered to ensure proper detection
-      setTimeout(() => {
-        this.setupScrollObserver();
-
-        // Force show the loading indicator if there are more words
-        if (this.wordsPagination.hasMore) {
-          this.showLoadingIndicator();
-        }
-      }, 100);
-
-    } catch (error) {
-      console.error('Error loading words:', error);
-    }
+    await this.wordManager.loadWords();
   }
 
-  private async loadMoreWords(): Promise<void> {
-    if (this.wordsPagination.isLoading || !this.wordsPagination.hasMore) return;
 
-    this.wordsPagination.isLoading = true;
-    this.showLoadingIndicator();
-
-    try {
-      const result = await window.electronAPI.getWordsPaginated(
-        this.wordsPagination.offset,
-        this.wordsPagination.pageSize,
-        this.sortOrder
-      );
-
-      // Filter out duplicates based on word ID
-      const existingIds = new Set(this.words.map(word => word.id));
-      const newWords = result.words.filter(word => !existingIds.has(word.id));
-
-      // Add new words to our collection (only non-duplicates)
-      this.words.push(...newWords);
-      this.wordsPagination.hasMore = result.hasMore;
-      this.wordsPagination.total = result.total;
-      this.wordsPagination.offset += this.wordsPagination.pageSize;
-
-      // Render the new words (only non-duplicates)
-      if (newWords.length > 0) {
-        this.wordRenderer.renderWordListIncremental(newWords);
-      }
-
-      // Update word count display
-      this.uiUtils.updateWordCount(this.wordsPagination.total);
-    } catch (error) {
-      console.error('Error loading more words:', error);
-      this.toastManager.showError('Failed to load more words');
-    } finally {
-      this.wordsPagination.isLoading = false;
-      this.updateLoadingIndicator();
-    }
-  }
-
-  private setupScrollObserver(): void {
-    // Create loading indicator element if it doesn't exist
-    let loadingIndicator = document.getElementById('loading-indicator');
-    if (!loadingIndicator) {
-      loadingIndicator = document.createElement('div');
-      loadingIndicator.id = 'loading-indicator';
-      loadingIndicator.className = 'flex justify-center items-center py-4 text-slate-500';
-      loadingIndicator.innerHTML = `
-        <div class="animate-spin rounded-full h-6 w-6 border-b-2 border-slate-500"></div>
-        <span class="ml-2 text-sm">Loading more words...</span>
-      `;
-      loadingIndicator.style.display = 'none';
-
-      const wordList = document.getElementById('word-list')!;
-      wordList.appendChild(loadingIndicator);
-    }
-
-    // Setup intersection observer with robust duplicate prevention
-    let isTriggering = false;
-
-    this.scrollObserver = new IntersectionObserver(
-      (entries) => {
-        const entry = entries[0];
-
-        if (entry.isIntersecting && !isTriggering && this.wordsPagination.hasMore && !this.wordsPagination.isLoading) {
-          isTriggering = true;
-          this.loadMoreWords().finally(() => {
-            isTriggering = false;
-          });
-        }
-      },
-      {
-        root: document.getElementById('word-list'),
-        rootMargin: '0px 0px 0px 0px',
-        threshold: 0.1
-      }
-    );
-
-    this.scrollObserver.observe(loadingIndicator);
-  }
-
-  private showLoadingIndicator(): void {
-    const loadingIndicator = document.getElementById('loading-indicator');
-    if (loadingIndicator) {
-      loadingIndicator.style.display = 'flex';
-    }
-  }
-
-  private updateLoadingIndicator(): void {
-    const loadingIndicator = document.getElementById('loading-indicator');
-    if (!loadingIndicator) return;
-
-    if (this.wordsPagination.isLoading) {
-      loadingIndicator.innerHTML = `
-        <div class="animate-spin rounded-full h-6 w-6 border-b-2 border-slate-500"></div>
-        <span class="ml-2 text-sm text-slate-500">Loading more words...</span>
-      `;
-    } else if (!this.wordsPagination.hasMore) {
-      loadingIndicator.innerHTML = `
-        <div class="rounded-full h-6 w-6 border-2 border-slate-300 flex items-center justify-center">
-          <svg class="w-4 h-4 text-slate-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M5 13l4 4L19 7"></path>
-          </svg>
-        </div>
-        <span class="ml-2 text-sm text-slate-400">No more words to load</span>
-      `;
-    } else {
-      loadingIndicator.innerHTML = `
-        <div class="rounded-full h-6 w-6 border-2 border-slate-400 flex items-center justify-center">
-          <svg class="w-4 h-4 text-slate-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 14l-7 7m0 0l-7-7m7 7V3"></path>
-          </svg>
-        </div>
-        <span class="ml-2 text-sm text-slate-400">Scroll for more words</span>
-      `;
-    }
-  }
 
   private setupEventListeners(): void {
-    // Set up word meaning streaming listener
-    window.electronAPI.onWordMeaningStreaming((content: string) => {
-      this.handleWordMeaningStreaming(content);
-    });
-
-    // Set up word metadata ready listener
-    window.electronAPI.onWordMetadataReady((wordMeta: any) => {
-      this.handleWordMetadataReady(wordMeta);
-    });
-
-    // Set up protocol handlers for custom URL scheme
-    window.electronAPI.onProtocolNavigateWord((wordName: string) => {
-      console.log('🎯 Renderer: Received protocol navigation request for word:', wordName);
-      this.handleProtocolNavigateWord(wordName);
-    });
-
-    window.electronAPI.onProtocolSwitchProfile((profileName: string) => {
-      console.log('🎯 Renderer: Received protocol profile switch request for:', profileName);
-      this.handleProtocolSwitchProfile(profileName);
-    });
-
-    // Profile selector
-    const profileSelect = document.getElementById('profile-select') as HTMLSelectElement;
-    profileSelect.addEventListener('change', async (e) => {
-      const target = e.target as HTMLSelectElement;
-      const newProfile = target.value;
-      if (newProfile && newProfile !== this.profileService.getCurrentProfile()) {
-        // Show loading overlay
-        this.showLoadingOverlay();
-
-        // Disable profile select during switching
-        profileSelect.disabled = true;
-
-        try {
-          const success = await this.profileService.switchProfile(newProfile);
-          if (success) {
-            // Successful switch - load words normally
-            this.resetUIForProfileSwitch();
-            await this.loadWords();
-          } else {
-            // Failed switch - but still switch UI to failed profile so user can delete it
-            this.profileService.setCurrentProfile(newProfile); // Update UI state
-            profileSelect.value = newProfile; // Keep dropdown on failed profile
-            this.resetUIForProfileSwitch(); // Reset UI
-            // Don't load words - just show empty state
-
-            this.toastManager.showError('Failed to load profile data. You can delete this profile in Settings.');
-          }
-        } catch (error) {
-          console.error('Error switching profile:', error);
-          // Even on exception, switch UI to failed profile so user can delete it
-          this.profileService.setCurrentProfile(newProfile);
-          profileSelect.value = newProfile;
-          this.resetUIForProfileSwitch();
-          this.toastManager.showError('Failed to load profile data. You can delete this profile in Settings.');
-        } finally {
-          setTimeout(() => {
-              this.hideLoadingOverlay();
-              profileSelect.disabled = false;
-            }, 500);
-        }
-      }
-    });
-
-    // Word input
-    const wordInput = document.getElementById('word-input') as HTMLInputElement;
-    wordInput.addEventListener('input', (e) => {
-      const target = e.target as HTMLInputElement;
-      const query = target.value.trim();
-      this.updateGenerateBtnState(query);
-      if (query.length > 0) {
-        this.handleSearchInput(query);
-      } else {
-        this.uiUtils.clearSuggestions();
-      }
-    });
-
-    // Hide suggestions when clicking outside
-    document.addEventListener('click', (e) => {
-      const target = e.target as HTMLElement;
-      const suggestionsDiv = document.getElementById('suggestions')!;
-      const wordInput = document.getElementById('word-input')!;
-
-      if (!suggestionsDiv.contains(target) && target !== wordInput) {
-        this.uiUtils.hideSuggestions();
-      }
-    });
-
-    // Hide suggestions when input loses focus
-    wordInput.addEventListener('blur', () => {
-      setTimeout(() => {
-        this.uiUtils.hideSuggestions();
-      }, 150);
-    });
-
-    // Generate button
-    const generateBtn = document.getElementById('generate-btn') as HTMLButtonElement;
-    generateBtn.addEventListener('click', () => {
-      if (this.isSearchMode) {
-        this.handleSearchExistingWord();
-      } else {
-        this.handleGenerate();
-      }
-    });
-
-    // Initially disable the generate button
-    generateBtn.disabled = true;
-
-    // Enter key on input
-    wordInput.addEventListener('keypress', (e) => {
-      if (e.key === 'Enter') {
-        const generateBtn = document.getElementById('generate-btn') as HTMLButtonElement;
-        if (this.isGenerating || (generateBtn && generateBtn.disabled)) {
-          return;
-        }
-
-        if (this.isSearchMode) {
-          this.handleSearchExistingWord();
-        } else {
-          this.handleGenerate();
-        }
-      }
-    });
-
-    // More button and inline actions
-    const moreBtn = document.getElementById('more-btn') as HTMLButtonElement;
-    const inlineActions = document.getElementById('inline-actions') as HTMLElement;
-
-    moreBtn.addEventListener('click', (e) => {
-      e.stopPropagation();
-      this.uiUtils.toggleInlineActions();
-    });
-
-    // Add profile button
-    const addProfileBtn = document.getElementById('add-profile-btn') as HTMLButtonElement;
-    addProfileBtn.addEventListener('click', () => {
-      this.showAddProfileModal();
-    });
-
-    // Export button
-    const exportBtn = document.getElementById('export-profile-btn') as HTMLButtonElement;
-    exportBtn.addEventListener('click', () => {
-      this.showExportChoiceModal();
-    });
-
-    // Import button
-    const importBtn = document.getElementById('import-profile-btn') as HTMLButtonElement;
-    importBtn.addEventListener('click', () => {
-      this.showImportChoiceModal();
-    });
-
-    // Import words button
-    const importWordsBtn = document.getElementById('import-words-btn') as HTMLButtonElement;
-    if (importWordsBtn) {
-      importWordsBtn.addEventListener('click', () => {
-        this.showImportWordsModal();
-      });
-    }
-
-    // Import words modal buttons
-    const selectFileBtn = document.getElementById('select-import-file') as HTMLButtonElement;
-    const startImportBtn = document.getElementById('start-import-btn') as HTMLButtonElement;
-    const cancelImportBtn = document.getElementById('cancel-import-btn') as HTMLButtonElement;
-    const closeImportModalBtn = document.getElementById('close-import-modal') as HTMLButtonElement;
-
-    if (selectFileBtn) {
-      selectFileBtn.addEventListener('click', () => {
-        this.selectImportFile();
-      });
-    }
-
-    if (startImportBtn) {
-      startImportBtn.addEventListener('click', () => {
-        this.startWordImport();
-      });
-    }
-
-    if (cancelImportBtn) {
-      cancelImportBtn.addEventListener('click', () => {
-        this.cancelWordImport();
-      });
-    }
-
-    if (closeImportModalBtn) {
-      closeImportModalBtn.addEventListener('click', () => {
-        this.hideImportWordsModal();
-      });
-    }
-
-    // Howto button
-    const howtoBtn = document.getElementById('howto-btn') as HTMLButtonElement;
-    howtoBtn.addEventListener('click', () => {
-      this.showHowtoModal();
-    });
-
-    // Close howto modal
-    const closeHowtoBtn = document.getElementById('close-howto-btn') as HTMLButtonElement;
-    closeHowtoBtn.addEventListener('click', () => this.hideHowtoModal());
-
-    // Add profile modal buttons
-    const cancelAddProfileBtn = document.getElementById('cancel-add-profile') as HTMLButtonElement;
-    const createProfileBtn = document.getElementById('create-profile') as HTMLButtonElement;
-
-    cancelAddProfileBtn.addEventListener('click', () => this.hideAddProfileModal());
-    createProfileBtn.addEventListener('click', () => this.handleCreateProfile());
-
-    // Settings button
-    const settingsBtn = document.getElementById('settings-btn') as HTMLButtonElement;
-    settingsBtn.addEventListener('click', () => this.showSettingsModal());
-
-    // Settings modal buttons
-    const cancelSettingsBtn = document.getElementById('cancel-settings') as HTMLButtonElement;
-    const saveSettingsBtn = document.getElementById('save-settings') as HTMLButtonElement;
-    const deleteProfileBtn = document.getElementById('delete-profile-btn') as HTMLButtonElement;
-    const toggleApiKeyBtn = document.getElementById('toggle-api-key-visibility') as HTMLButtonElement;
-
-    cancelSettingsBtn.addEventListener('click', () => this.hideSettingsModal());
-    saveSettingsBtn.addEventListener('click', () => this.saveSettings());
-    deleteProfileBtn.addEventListener('click', () => this.handleDeleteProfile());
-    toggleApiKeyBtn.addEventListener('click', () => this.toggleApiKeyVisibility());
-
-    // Sort button
-    const sortBtn = document.getElementById('sort-btn') as HTMLButtonElement;
-    if (sortBtn) {
-      sortBtn.addEventListener('click', () => this.handleSortToggle());
-    }
-
-    // Resize functionality
-    const mainContent = document.getElementById('main-content') as HTMLElement;
-    mainContent.addEventListener('mousedown', (e) => this.uiUtils.startResize(e));
-
-    // Global mouse events for resizing
-    document.addEventListener('mousemove', (e) => this.uiUtils.handleResize(e));
-    document.addEventListener('mouseup', () => this.uiUtils.stopResize());
-
-    // Google Drive modal event handlers
-    this.setupGoogleDriveModalHandlers();
+    // Delegate event setup to EventManager
+    this.eventManager.setupEventListeners();
   }
 
-  private setupGoogleDriveModalHandlers(): void {
-    // Export choice modal buttons
-    const exportLocalBtn = document.getElementById('export-local-btn') as HTMLButtonElement;
-    const exportGoogleDriveBtn = document.getElementById('export-google-drive-btn') as HTMLButtonElement;
-    const cancelExportChoice = document.getElementById('cancel-export-choice') as HTMLButtonElement;
-     // Google Drive upload success modal
-    const closeUploadModalBtn = document.getElementById('close-google-drive-upload-modal') as HTMLButtonElement;
-    
-    if (closeUploadModalBtn) {
-      closeUploadModalBtn.addEventListener('click', () => {
-        this.hideGoogleDriveUploadModal();
-      });
-    }
 
-    if (exportLocalBtn) {
-      exportLocalBtn.addEventListener('click', () => {
-        this.hideExportChoiceModal();
-        this.handleExportProfile();
-      });
-    }
-
-    if (exportGoogleDriveBtn) {
-      exportGoogleDriveBtn.addEventListener('click', async () => {
-        this.hideExportChoiceModal();
-        await this.handleExportToGoogleDrive();
-      });
-    }
-
-    if (cancelExportChoice) {
-      cancelExportChoice.addEventListener('click', () => {
-        this.hideExportChoiceModal();
-      });
-    }
-
-    // Import choice modal buttons
-    const importLocalBtn = document.getElementById('import-local-btn') as HTMLButtonElement;
-    const importGoogleDriveBtn = document.getElementById('import-google-drive-btn') as HTMLButtonElement;
-    const cancelImportChoice = document.getElementById('cancel-import-choice') as HTMLButtonElement;
-
-    if (importLocalBtn) {
-      importLocalBtn.addEventListener('click', () => {
-        this.hideImportChoiceModal();
-        this.handleImportProfile();
-      });
-    }
-
-    if (importGoogleDriveBtn) {
-      importGoogleDriveBtn.addEventListener('click', async () => {
-        this.hideImportChoiceModal();
-        await this.showGoogleDriveFilePicker();
-      });
-    }
-
-    if (cancelImportChoice) {
-      cancelImportChoice.addEventListener('click', () => {
-        this.hideImportChoiceModal();
-      });
-    }
-
-    // Google Drive file picker modal
-    const cancelGoogleDrivePicker = document.getElementById('cancel-google-drive-picker') as HTMLButtonElement;
-    const importGoogleDriveFile = document.getElementById('import-google-drive-file') as HTMLButtonElement;
-
-    if (cancelGoogleDrivePicker) {
-      cancelGoogleDrivePicker.addEventListener('click', () => {
-        this.hideGoogleDriveFilePicker();
-        this.selectedGoogleDriveFile = null;
-        this.updateGoogleDriveImportUI();
-      });
-    }
-
-    if (importGoogleDriveFile) {
-      importGoogleDriveFile.addEventListener('click', async () => {
-        if (this.selectedGoogleDriveFile) {
-          await this.performGoogleDriveImport(this.selectedGoogleDriveFile.id);
-        }
-      });
-    }
-  }
-
-  private async handleSearchInput(query: string): Promise<void> {
-    try {
-      const suggestions = await this.wordService.searchWords(query);
-      this.wordRenderer.renderSuggestions(suggestions);
-
-      const hasExactMatch = suggestions.some((word: WordListItem) =>
-        word.word.toLowerCase() === query.toLowerCase()
-      );
-      this.updateGenerateBtnState(query, hasExactMatch);
-    } catch (error) {
-      console.error('Error searching words:', error);
-    }
-  }
 
   private async handleGenerate(): Promise<void> {
-    const wordInput = document.getElementById('word-input') as HTMLInputElement;
-    const word = wordInput.value.trim();
-
-    if (!word) {
-      this.toastManager.showError('Please enter a word');
-      return;
-    }
-
-    const generateBtn = document.getElementById('generate-btn') as HTMLButtonElement;
-    const generateIcon = document.getElementById('generate-icon') as unknown as SVGElement;
-    const loadingIcon = document.getElementById('loading-icon') as unknown as SVGElement;
-
-    const generationId = generateGenerationId();
-    console.log('🔄 Renderer: Generated new generationId:', generationId);
-
-    if (!generationId || typeof generationId !== 'string' || generationId.length === 0) {
-      console.error('❌ Renderer: Invalid generationId generated:', generationId);
-      this.toastManager.showError('Failed to generate request ID. Please try again.');
-      return;
-    }
-
-    this.currentGenerationId = generationId;
-    this.isGenerating = true;
-    this.wordRenderer.setGenerationState(true); // Update WordRenderer generation state
-    generateBtn.disabled = true;
-    generateIcon.classList.add('hidden');
-    loadingIcon.classList.remove('hidden');
-
-    try {
-      this.uiUtils.clearWordDetails();
-      this.streamingContent = '';
-
-      const tempWord: WordDocument = {
-        id: 'temp',
-        word,
-        one_line_desc: '',
-        details: '',
-        tags: [],
-        tag_colors: {},
-        synonyms: [],
-        antonyms: [],
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString()
-      };
-
-      this.currentWord = tempWord;
-      await this.wordRenderer.renderStreamingWordDetails(tempWord);
-
-      const meaning = await this.wordService.generateWordMeaning(word);
-      tempWord.details = meaning;
-      this.currentWord = tempWord;
-
-      tempWord.one_line_desc = 'Generating summary...';
-      tempWord.tags = ['Analyzing word...'];
-      tempWord.tag_colors = { 'Analyzing word...': '#6b7280' };
-      await this.wordRenderer.renderWordDetails(tempWord);
-
-      setTimeout(() => {
-        const wordDetails = document.getElementById('word-details');
-        if (wordDetails && wordDetails.parentElement) {
-          wordDetails.parentElement.scrollTop = wordDetails.parentElement.scrollHeight;
-        }
-      }, 100);
-
-      try {
-        console.log('Starting generateWordMetas call...');
-        const tagsResult = await this.wordService.generateWordMetas(word, meaning, generationId);
-        console.log('generateWordMetas completed:', tagsResult);
-
-        setTimeout(() => {
-          if (this.currentGenerationId === generationId &&
-            this.currentWord &&
-            this.currentWord.tags.includes('Analyzing word...')) {
-            console.log('Tool result timeout - forcing UI update');
-            this.currentWord.one_line_desc = `Summary for: ${word}`;
-            this.currentWord.tags = ['general'];
-            this.currentWord.tag_colors = { 'general': '#6b7280' };
-            this.wordRenderer.renderWordDetails(this.currentWord);
-          }
-        }, 10000);
-
-      } catch (error) {
-        console.error('Error in generateWordMetas:', error);
-        this.toastManager.showError('Failed to generate tags and summary. Please check your API configuration.');
-        if (this.currentGenerationId === generationId && this.currentWord) {
-          this.currentWord.one_line_desc = 'Failed to generate summary';
-          this.currentWord.tags = ['Failed to generate tags'];
-          this.currentWord.tag_colors = { 'Failed to generate tags': '#ef4444' };
-          this.wordRenderer.renderWordDetails(this.currentWord);
-        }
-      }
-
-    } catch (error) {
-      console.error('Error generating meaning:', error);
-      this.toastManager.showError('Failed to generate meaning. Please check your API configuration.');
-    } finally {
-      this.isGenerating = false;
-      this.wordRenderer.setGenerationState(false); // Update WordRenderer generation state
-      generateBtn.disabled = false;
-      generateIcon.classList.remove('hidden');
-      loadingIcon.classList.add('hidden');
-    }
+    await this.wordManager.handleGenerate();
   }
 
-  private async handleAddWord(word: WordDocument): Promise<void> {
-    if (!this.currentWord) {
-      this.toastManager.showError('No word to add');
-      return;
-    }
 
-    try {
-      const wordData = {
-        word: this.currentWord.word,
-        one_line_desc: this.currentWord.one_line_desc,
-        details: this.currentWord.details,
-        tags: this.currentWord.tags,
-        tag_colors: this.currentWord.tag_colors,
-        synonyms: this.currentWord.synonyms || [],
-        antonyms: this.currentWord.antonyms || []
-      };
-
-      const addedWord = await this.wordService.addWord(wordData);
-      this.currentWord = addedWord;
-
-      this.words.unshift({
-        id: addedWord.id,
-        word: addedWord.word,
-        one_line_desc: addedWord.one_line_desc
-      });
-
-      this.wordsPagination.total++;
-      this.wordRenderer.renderWordItemAtTop(addedWord);
-      this.uiUtils.updateWordCount(this.wordsPagination.total);
-
-      if (this.currentWord) {
-        await this.wordRenderer.renderWordDetails(this.currentWord);
-      }
-
-      this.updateGenerateBtnState(addedWord.word, true);
-      this.toastManager.showSuccess('Word added successfully');
-    } catch (error) {
-      console.error('Error adding word:', error);
-      this.toastManager.showError('Failed to add word');
-    }
-  }
-
-  private async handleRefreshWord(word: WordDocument): Promise<void> {
-    if (!this.currentWord || this.currentWord.id === 'temp') {
-      this.toastManager.showError('No word selected');
-      return;
-    }
-
-    const originalWordId = this.currentWord.id;
-    const originalCreatedAt = this.currentWord.created_at;
-
-    try {
-      const wordInput = document.getElementById('word-input') as HTMLInputElement;
-      wordInput.value = this.currentWord.word;
-
-      await this.handleGenerate();
-
-      if (this.currentWord && this.currentWord.id === 'temp') {
-        const updatedWordData = {
-          word: this.currentWord.word,
-          one_line_desc: this.currentWord.one_line_desc,
-          details: this.currentWord.details,
-          tags: this.currentWord.tags,
-          tag_colors: this.currentWord.tag_colors,
-          synonyms: this.currentWord.synonyms || [],
-          antonyms: this.currentWord.antonyms || []
-        };
-
-        const updatedWord = await this.wordService.updateWord(originalWordId, updatedWordData);
-
-        if (updatedWord) {
-          updatedWord.id = originalWordId;
-          updatedWord.created_at = originalCreatedAt;
-          updatedWord.updated_at = new Date().toISOString();
-
-          this.currentWord = updatedWord;
-
-          const wordIndex = this.words.findIndex(word => word.id === originalWordId);
-          if (wordIndex !== -1) {
-            Object.assign(this.words[wordIndex], {
-              id: updatedWord.id,
-              word: updatedWord.word,
-              one_line_desc: updatedWord.one_line_desc
-            });
-          }
-
-          if (this.currentWord) {
-            await this.wordRenderer.renderWordDetails(this.currentWord);
-            this.uiUtils.updateWordInList(originalWordId, this.currentWord);
-          }
-        } else {
-          this.toastManager.showError('Failed to update word');
-        }
-
-        this.toastManager.showSuccess('Word refreshed successfully');
-      }
-    } catch (error) {
-      console.error('Error refreshing word:', error);
-      this.toastManager.showError('Failed to refresh word');
-    }
-  }
-
-  private async handleDeleteWord(word: WordDocument): Promise<void> {
-    if (!this.currentWord || this.currentWord.id === 'temp') {
-      this.toastManager.showError('No word to delete');
-      return;
-    }
-
-    if (!confirm(`Are you sure you want to delete "${this.currentWord.word}"?`)) {
-      return;
-    }
-
-    try {
-      const success = await this.wordService.deleteWord(this.currentWord.id);
-      if (success) {
-        const deletedWordId = this.currentWord.id;
-
-        this.words = this.words.filter(word => word.id !== deletedWordId);
-
-        const wordItem = document.querySelector(`[data-word-id="${deletedWordId}"]`) as HTMLElement;
-        if (wordItem) {
-          wordItem.remove();
-        }
-
-        this.wordsPagination.total = Math.max(0, this.wordsPagination.total - 1);
-        this.uiUtils.updateWordCount(this.wordsPagination.total);
-
-        this.currentWord = null;
-        this.uiUtils.clearWordDetails();
-        this.updateLoadingIndicator();
-
-        this.toastManager.showSuccess('Word deleted successfully');
-      } else {
-        this.toastManager.showError('Failed to delete word');
-      }
-    } catch (error) {
-      console.error('Error deleting word:', error);
-      this.toastManager.showError('Failed to delete word');
-    }
-  }
 
   private async loadAssociatedWords(tag: string): Promise<void> {
-    if (this.associatedWordsState.scrollObserver) {
-      this.associatedWordsState.scrollObserver.disconnect();
-      this.associatedWordsState.scrollObserver = null;
-    }
-
-    this.associatedWordsState = {
-      words: [],
-      offset: 0,
-      pageSize: ASSOCIATED_WORDS_PAGE_SIZE,
-      isLoading: false,
-      hasMore: true,
-      total: 0,
-      currentTag: tag,
-      scrollObserver: null
-    };
-
-    const associatedList = document.getElementById('associated-list')!;
-    associatedList.innerHTML = '';
-
-    await this.loadMoreAssociatedWords();
-
-    setTimeout(() => {
-      this.setupAssociatedScrollObserver();
-
-      if (this.associatedWordsState.hasMore) {
-        this.showAssociatedLoadingIndicator();
-      }
-    }, 100);
-  }
-
-  private async loadMoreAssociatedWords(): Promise<void> {
-    if (this.associatedWordsState.isLoading || !this.associatedWordsState.hasMore) {
-      return;
-    }
-
-    this.associatedWordsState.isLoading = true;
-    this.showAssociatedLoadingIndicator();
-
-    try {
-      const result = await window.electronAPI.getRelatedWordsPaginated(
-        this.associatedWordsState.currentTag,
-        this.associatedWordsState.offset,
-        this.associatedWordsState.pageSize
-      );
-
-      const existingIds = new Set(this.associatedWordsState.words.map(word => word.id));
-      const newWords = result.words.filter(word => !existingIds.has(word.id));
-
-      this.associatedWordsState.words.push(...newWords);
-      this.associatedWordsState.hasMore = result.hasMore;
-      this.associatedWordsState.offset += this.associatedWordsState.pageSize;
-
-      if (newWords.length > 0) {
-        this.wordRenderer.renderAssociatedWordListIncremental(newWords);
-      }
-
-      this.uiUtils.updateAssociatedCount(result.total);
-    } catch (error) {
-      console.error('Error loading more associated words:', error);
-      this.toastManager.showError('Failed to load more associated words');
-    } finally {
-      this.associatedWordsState.isLoading = false;
-      this.updateAssociatedLoadingIndicator();
-    }
-  }
-
-  private setupAssociatedScrollObserver(): void {
-    let loadingIndicator = document.getElementById('associated-loading-indicator');
-    if (!loadingIndicator) {
-      loadingIndicator = document.createElement('div');
-      loadingIndicator.id = 'associated-loading-indicator';
-      loadingIndicator.className = 'flex justify-center items-center py-4 text-slate-500';
-      loadingIndicator.innerHTML = `
-        <div class="animate-spin rounded-full h-6 w-6 border-b-2 border-slate-500"></div>
-        <span class="ml-2 text-sm">Loading more words...</span>
-      `;
-      loadingIndicator.style.display = 'none';
-
-      const associatedList = document.getElementById('associated-list')!;
-      associatedList.appendChild(loadingIndicator);
-    }
-
-    let isTriggering = false;
-
-    this.associatedWordsState.scrollObserver = new IntersectionObserver(
-      (entries) => {
-        const entry = entries[0];
-
-        if (entry.isIntersecting && !isTriggering && this.associatedWordsState.hasMore && !this.associatedWordsState.isLoading) {
-          isTriggering = true;
-          this.loadMoreAssociatedWords().finally(() => {
-            isTriggering = false;
-          });
-        }
-      },
-      {
-        root: document.getElementById('associated-list'),
-        rootMargin: '0px 0px 0px 0px',
-        threshold: 0.1
-      }
-    );
-
-    this.associatedWordsState.scrollObserver.observe(loadingIndicator);
-  }
-
-  private showAssociatedLoadingIndicator(): void {
-    const loadingIndicator = document.getElementById('associated-loading-indicator');
-    if (loadingIndicator) {
-      loadingIndicator.style.display = 'flex';
-    }
-  }
-
-  private updateAssociatedLoadingIndicator(): void {
-    const loadingIndicator = document.getElementById('associated-loading-indicator');
-    if (!loadingIndicator) return;
-
-    if (this.associatedWordsState.isLoading) {
-      loadingIndicator.innerHTML = `
-        <div class="animate-spin rounded-full h-6 w-6 border-b-2 border-slate-500"></div>
-        <span class="ml-2 text-sm text-slate-500">Loading more words...</span>
-      `;
-    } else if (!this.associatedWordsState.hasMore) {
-      loadingIndicator.innerHTML = `
-        <div class="rounded-full h-6 w-6 border-2 border-slate-300 flex items-center justify-center">
-          <svg class="w-4 h-4 text-slate-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M5 13l4 4L19 7"></path>
-          </svg>
-        </div>
-        <span class="ml-2 text-sm text-slate-400">No more words to load</span>
-      `;
-    } else {
-      loadingIndicator.innerHTML = `
-        <div class="rounded-full h-6 w-6 border-2 border-slate-400 flex items-center justify-center">
-          <svg class="w-4 h-4 text-slate-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 14l-7 7m0 0l-7-7m7 7V3"></path>
-          </svg>
-        </div>
-        <span class="ml-2 text-sm text-slate-400">Scroll for more words</span>
-      `;
-    }
+    await this.associatedWordsManager.loadAssociatedWords(tag);
   }
 
   private async selectWord(word: WordDocument | WordListItem): Promise<void> {
-    if (this.isGenerating) {
-      console.log('⚠️ selectWord: Generation in progress, ignoring word selection');
-      return;
-    }
-
-    let fullWord: WordDocument;
-
-    if ('details' in word && 'tags' in word && 'tag_colors' in word) {
-      fullWord = word as WordDocument;
-    } else {
-      try {
-        const fetchedWord = await this.wordService.getWord(word.id);
-        if (!fetchedWord) {
-          console.error('Failed to fetch full word details for:', word.id);
-          this.toastManager.showError('Failed to load word details');
-          return;
-        }
-        fullWord = fetchedWord;
-      } catch (error) {
-        console.error('Error fetching word details:', error);
-        this.toastManager.showError('Failed to load word details');
-        return;
-      }
-    }
-
-    this.currentWord = fullWord;
-    await this.wordRenderer.renderWordDetails(fullWord);
-
-    const wordInput = document.getElementById('word-input') as HTMLInputElement;
-    wordInput.value = fullWord.word;
-
-    this.uiUtils.clearSuggestions();
-    this.updateGenerateBtnState(fullWord.word, true);
+    await this.wordManager.selectWord(word);
   }
 
-  private resetUIForProfileSwitch(): void {
-    this.currentWord = null;
-    this.uiUtils.clearWordDetails();
 
-    const wordInput = document.getElementById('word-input') as HTMLInputElement;
-    if (wordInput) {
-      wordInput.value = '';
-    }
 
-    this.uiUtils.clearSuggestions();
 
-    this.isGenerating = false;
-    this.wordRenderer.setGenerationState(false); // Reset WordRenderer generation state
-    this.currentGenerationId = '';
-    this.streamingContent = '';
 
-    const wordList = document.getElementById('word-list')!;
-    if (wordList) {
-      wordList.innerHTML = '';
-    }
 
-    const associatedList = document.getElementById('associated-list')!;
-    if (associatedList) {
-      associatedList.innerHTML = '';
-    }
-
-    this.words = [];
-    this.wordsPagination = {
-      offset: 0,
-      pageSize: WORDS_PAGE_SIZE,
-      isLoading: false,
-      hasMore: true,
-      total: 0
-    };
-
-    this.associatedWordsState = {
-      words: [],
-      offset: 0,
-      pageSize: ASSOCIATED_WORDS_PAGE_SIZE,
-      isLoading: false,
-      hasMore: true,
-      total: 0,
-      currentTag: '',
-      scrollObserver: null
-    };
-
-    if (this.scrollObserver) {
-      this.scrollObserver.disconnect();
-      this.scrollObserver = null;
-    }
-    if (this.associatedWordsState.scrollObserver) {
-      this.associatedWordsState.scrollObserver.disconnect();
-      this.associatedWordsState.scrollObserver = null;
-    }
-
-    this.updateGenerateBtnState('', false);
-    this.isSearchMode = false;
-
-    this.uiUtils.updateWordCount(0);
-    this.uiUtils.updateAssociatedCount(0);
-
-    // Reload panel widths for the new profile
-    this.uiUtils.loadPanelWidths();
-  }
-
-  private handleWordMeaningStreaming(content: string): void {
-    this.streamingContent += content;
-    if (this.currentWord && this.currentWord.id === 'temp') {
-      const streamingWord = { ...this.currentWord, details: this.streamingContent };
-      this.wordRenderer.renderStreamingWordDetails(streamingWord);
-
-      const wordDetails = document.getElementById('word-details');
-      if (wordDetails && wordDetails.parentElement) {
-        wordDetails.parentElement.scrollTop = wordDetails.parentElement.scrollHeight;
-      }
-    }
-  }
-
-  private handleWordMetadataReady(wordMeta: any): void {
-    // Simple check: if we have metadata and a current word, update it
-    if (!wordMeta || !this.currentWord || this.currentWord.id !== 'temp') {
-      return;
-    }
-
-    // Update word metadata
-    if (wordMeta.summary) {
-      this.currentWord.one_line_desc = wordMeta.summary;
-    }
-    if (wordMeta.tags) {
-      this.currentWord.tags = wordMeta.tags;
-    }
-    if (wordMeta.tag_colors) {
-      this.currentWord.tag_colors = wordMeta.tag_colors;
-    }
-    if (wordMeta.synonyms) {
-      this.currentWord.synonyms = wordMeta.synonyms;
-    }
-    if (wordMeta.antonyms) {
-      this.currentWord.antonyms = wordMeta.antonyms;
-    }
-
-    // Re-render word details
-    this.wordRenderer.renderWordDetails(this.currentWord);
-  }
-
-  private updateGenerateBtnState(query: string, hasExactMatch?: boolean): void {
-    const generateBtn = document.getElementById('generate-btn') as HTMLButtonElement;
-    const generateIcon = document.getElementById('generate-icon') as unknown as SVGElement;
-
-    if (!generateBtn || !generateIcon) return;
-
-    generateBtn.disabled = query.length === 0;
-
-    if (hasExactMatch !== undefined) {
-      this.isSearchMode = hasExactMatch;
-
-      if (hasExactMatch) {
-        generateBtn.title = 'View Existing Word';
-        generateIcon.innerHTML = `
-          <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"></path>
-        `;
-      } else {
-        generateBtn.title = 'Generate';
-        generateIcon.innerHTML = `
-          <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 19l9 2-9-18-9 18 9-2zm0 0v-8"></path>
-        `;
-      }
-    }
-  }
 
   private async handleSearchExistingWord(): Promise<void> {
-    const wordInput = document.getElementById('word-input') as HTMLInputElement;
-    const query = wordInput.value.trim().toLowerCase();
-
-    if (!query) {
-      this.toastManager.showError('Please enter a word');
-      return;
-    }
-
-    try {
-      const suggestions = await this.wordService.searchWords(query);
-      const exactMatch = suggestions.find(word =>
-        word.word.toLowerCase() === query
-      );
-
-      if (exactMatch) {
-        this.selectWord(exactMatch);
-        this.uiUtils.clearSuggestions();
-      } else {
-        this.toastManager.showError('Word not found. Try generating it instead.');
-      }
-    } catch (error) {
-      console.error('Error searching for existing word:', error);
-      this.toastManager.showError('Failed to find existing word');
-    }
+    await this.wordManager.handleSearchExistingWord();
   }
 
   // Profile modal methods
   private showAddProfileModal(): void {
-    const input = document.getElementById('new-profile-name') as HTMLInputElement;
-    if (input) {
-      input.value = '';
-    }
-
-    const modal = document.getElementById('add-profile-modal')!;
-    modal.classList.remove('hidden');
-
-    setTimeout(() => {
-      if (input) {
-        input.focus();
-      }
-    }, 100);
+    this.modalManager.showAddProfileModal();
   }
 
   private hideAddProfileModal(): void {
-    const modal = document.getElementById('add-profile-modal')!;
-    modal.classList.add('hidden');
+    this.modalManager.hideAddProfileModal();
   }
 
   private async handleCreateProfile(): Promise<void> {
-    const input = document.getElementById('new-profile-name') as HTMLInputElement;
-    const profileName = input ? input.value.trim() : '';
-
-    if (!profileName) {
-      this.toastManager.showError('Profile name cannot be empty');
-      return;
-    }
-
-    if (this.profileService.getProfiles().includes(profileName)) {
-      this.toastManager.showError('A profile with this name already exists');
-      return;
-    }
-
-    try {
-      const success = await this.profileService.createProfile(profileName);
-      if (success) {
-        this.resetUIForProfileSwitch();
-        this.toastManager.showSuccess(`Profile "${profileName}" created successfully`);
-      } else {
-        this.toastManager.showError('Failed to create profile');
-      }
-    } catch (error) {
-      console.error('Error creating profile:', error);
-      this.toastManager.showError('Failed to create profile');
-    }
-
-    this.hideAddProfileModal();
+    await this.modalManager.handleCreateProfile();
   }
 
   private async handleDeleteProfile(): Promise<void> {
-    if (this.profileService.getProfiles().length <= 1) {
-      this.toastManager.showError('Cannot delete the last remaining profile');
-      return;
-    }
-
-    const confirmed = confirm(`Are you sure you want to delete the profile "${this.profileService.getCurrentProfile()}"? This action cannot be undone and will delete all words in this profile.`);
-    if (!confirmed) {
-      return;
-    }
-
-    try {
-      const success = await this.profileService.deleteProfile(this.profileService.getCurrentProfile());
-      if (success) {
-        this.resetUIForProfileSwitch();
-        await this.loadWords();
-        this.toastManager.showSuccess(`Profile "${this.profileService.getCurrentProfile()}" deleted successfully.`);
-      } else {
-        this.toastManager.showError('Failed to delete profile');
-      }
-    } catch (error) {
-      console.error('Error deleting profile:', error);
-      this.toastManager.showError('Failed to delete profile');
-    }
-
-    this.hideSettingsModal();
+    await this.modalManager.handleDeleteProfile();
   }
 
   private async handleExportProfile(): Promise<void> {
-    try {
-      const result = await this.profileService.exportProfile();
-      if (result.success) {
-        this.toastManager.showSuccess(result.message);
-      } else {
-        this.toastManager.showError(result.message);
-      }
-    } catch (error) {
-      console.error('Error exporting profile:', error);
-      this.toastManager.showError('Failed to export profile');
-    }
+    await this.modalManager.handleExportProfile();
   }
 
   private async handleImportProfile(): Promise<void> {
-    try {
-      const result = await this.profileService.importProfile();
-      if (result.success) {
-        this.toastManager.showSuccess(result.message);
-
-        if (result.profileName) {
-          // Refresh the profile list to include the new profile
-          await this.profileService.loadProfiles();
-
-          this.profileService.setCurrentProfile(result.profileName);
-          const profileSelect = document.getElementById('profile-select') as HTMLSelectElement;
-          profileSelect.value = result.profileName;
-
-          this.resetUIForProfileSwitch();
-          await this.loadWords();
-        }
-      } else {
-        this.toastManager.showError(result.message);
-      }
-    } catch (error) {
-      console.error('Error importing profile:', error);
-      this.toastManager.showError('Failed to import profile');
-    }
+    await this.modalManager.handleImportProfile();
   }
 
-  // Settings modal methods
+  // Settings modal methods - delegate to ModalManager
   private async showSettingsModal(): Promise<void> {
-    try {
-      const profileConfig = await this.profileService.getProfileConfig();
-      if (profileConfig) {
-        const profileNameInput = document.getElementById('profile-name') as HTMLInputElement;
-        if (profileNameInput) {
-          profileNameInput.value = this.profileService.getCurrentProfile() || '';
-        }
-        (document.getElementById('system-prompt') as HTMLTextAreaElement).value = profileConfig.system_prompt || '';
-        (document.getElementById('model-provider') as HTMLSelectElement).value = profileConfig.model_config.provider || 'openai';
-        (document.getElementById('model-name') as HTMLInputElement).value = profileConfig.model_config.model || '';
-        (document.getElementById('api-endpoint') as HTMLInputElement).value = profileConfig.model_config.endpoint || '';
-        (document.getElementById('api-key') as HTMLInputElement).value = profileConfig.model_config.api_key || '';
-      }
-    } catch (error) {
-      console.error('Error loading profile config:', error);
-      this.toastManager.showError('Failed to load profile settings');
-    }
-
-    const modal = document.getElementById('settings-modal')!;
-    modal.classList.remove('hidden');
+    await this.modalManager.showSettingsModal();
   }
 
   private hideSettingsModal(): void {
-    const modal = document.getElementById('settings-modal')!;
-    modal.classList.add('hidden');
+    this.modalManager.hideSettingsModal();
   }
 
   private async saveSettings(): Promise<void> {
-    const profileNameInput = document.getElementById('profile-name') as HTMLInputElement;
-    const profileName = profileNameInput ? profileNameInput.value.trim() : '';
-    const systemPrompt = (document.getElementById('system-prompt') as HTMLTextAreaElement).value;
-    const modelProvider = (document.getElementById('model-provider') as HTMLSelectElement).value;
-    const modelName = (document.getElementById('model-name') as HTMLInputElement).value;
-    const apiEndpoint = (document.getElementById('api-endpoint') as HTMLInputElement).value;
-    const apiKey = (document.getElementById('api-key') as HTMLInputElement).value;
-
-    if (!profileName) {
-      this.toastManager.showError('Profile name cannot be empty');
-      return;
-    }
-
-    const config = {
-      system_prompt: systemPrompt,
-      model_config: {
-        provider: modelProvider,
-        model: modelName,
-        endpoint: apiEndpoint,
-        api_key: apiKey
-      }
-    };
-
-    try {
-      const success = await this.profileService.updateProfileConfig(config);
-      if (success) {
-        if (profileName !== this.profileService.getCurrentProfile()) {
-          const success = await this.profileService.renameProfile(this.profileService.getCurrentProfile(), profileName);
-          if (success) {
-            await this.profileService.loadProfiles();
-            this.profileService.setCurrentProfile(profileName);
-            const profileSelect = document.getElementById('profile-select') as HTMLSelectElement;
-            profileSelect.value = profileName;
-          } else {
-            this.toastManager.showError('Failed to rename profile');
-            return;
-          }
-        }
-
-        this.toastManager.showSuccess('Settings saved successfully');
-        this.hideSettingsModal();
-      } else {
-        this.toastManager.showError('Failed to save settings');
-      }
-    } catch (error) {
-      console.error('Error saving settings:', error);
-      this.toastManager.showError('Failed to save settings');
-    }
+    await this.modalManager.saveSettings();
   }
 
-  // Howto modal methods
+  // Howto modal methods - delegate to ModalManager
   private async showHowtoModal(): Promise<void> {
-    try {
-      // Load markdown content from assets/howto.md
-      const response = await fetch('../../assets/howto.md');
-      if (!response.ok) {
-        throw new Error(`Failed to load howto.md: ${response.status}`);
-      }
-
-      const markdown = await response.text();
-
-      // Convert markdown to HTML using the existing IPC method
-      const htmlContent = await window.electronAPI.processMarkdown(markdown);
-
-      // Insert into modal
-      const contentDiv = document.getElementById('howto-content')!;
-      contentDiv.innerHTML = htmlContent;
-
-      // Show modal
-      const modal = document.getElementById('howto-modal')!;
-      modal.classList.remove('hidden');
-    } catch (error) {
-      console.error('Error loading howto content:', error);
-      this.toastManager.showError('Failed to load help content');
-    }
+    await this.modalManager.showHowtoModal();
   }
 
   private hideHowtoModal(): void {
-    const modal = document.getElementById('howto-modal')!;
-    modal.classList.add('hidden');
-  }
-
-  private async handleWriteRemark(word: WordDocument): Promise<void> {
-    if (!word || word.id === 'temp') {
-      this.toastManager.showError('Cannot add remark to temporary word');
-      return;
-    }
-
-    // Find the word details container
-    const wordDetails = document.getElementById('word-details');
-    if (!wordDetails) {
-      this.toastManager.showError('Word details not found');
-      return;
-    }
-
-    let remarkContainer: HTMLElement;
-    let remarkDisplay: HTMLElement | null = document.querySelector('.remark-display') as HTMLElement;
-
-    if (remarkDisplay) {
-      // Existing remark - replace the display
-      remarkContainer = remarkDisplay.parentElement!;
-    } else {
-      // No existing remark - create the remark section
-      remarkContainer = document.createElement('div');
-      remarkContainer.innerHTML = `
-        <h4 class="text-lg font-semibold text-slate-800 mb-3 flex items-center">
-          <svg class="w-4 h-4 mr-2 text-violet-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M11 5H6a2 2 0 00-2 2v11a2 2 0 002 2h11a2 2 0 002-2v-5m-1.414-9.414a2 2 0 112.828 2.828L11.828 15H9v-2.828l8.586-8.586z"></path>
-          </svg>
-          Remark
-        </h4>
-      `;
-
-      // Find the action buttons container and its parent (the content container)
-      const actionButtonsContainer = document.getElementById('action-buttons-container');
-      if (actionButtonsContainer) {
-        const contentContainer = actionButtonsContainer.parentNode as HTMLElement;
-        if (contentContainer) {
-          // Insert the remark section right before the action buttons
-          contentContainer.insertBefore(remarkContainer, actionButtonsContainer);
-        } else {
-          // Fallback: append to word details
-          wordDetails.appendChild(remarkContainer);
-        }
-      } else {
-        // Fallback: append to word details
-        wordDetails.appendChild(remarkContainer);
-      }
-    }
-
-    // Create input field
-    const input = document.createElement('input');
-    const originalRemark = word.remark || '';
-    input.type = 'text';
-    input.value = originalRemark;
-    input.className = 'w-full px-3 py-2 focus:outline-none focus:ring-0 text-slate-700';
-    input.style.background = 'transparent';
-    input.style.border = 'none';
-    input.style.outline = 'none';
-    input.style.boxShadow = 'none';
-    input.style.borderRadius = '0';
-    input.style.padding = '0.5rem 0.75rem';
-    input.style.color = '#334155';
-    input.placeholder = 'Enter your remark...';
-
-    if (remarkDisplay) {
-      // Replace existing display
-      remarkContainer.replaceChild(input, remarkDisplay);
-    } else {
-      // Add to new container
-      remarkContainer.appendChild(input);
-    }
-
-    // Focus and select all text
-    input.focus();
-    input.select();
-
-    let saved = false;
-    let cancelled = false;
-
-    const saveRemark = async () => {
-      if (saved) return;
-      saved = true;
-
-      const remarkValue = input.value.trim();
-
-      // Check if the remark has actually changed
-      if (remarkValue === originalRemark) {
-        // No change, just restore the display without saving
-        if (remarkDisplay) {
-          remarkContainer.replaceChild(remarkDisplay, input);
-        } else {
-          // If there was no original display, remove the input and the remark section
-          remarkContainer.remove();
-        }
-        return;
-      }
-
-      try {
-        // Update word in database
-        const updatedWord = await this.wordService.updateWord(word.id, { remark: remarkValue });
-
-        if (updatedWord) {
-          // Update current word
-          this.currentWord = { ...this.currentWord!, remark: remarkValue };
-
-          // Update word in list if it exists
-          const wordIndex = this.words.findIndex(w => w.id === word.id);
-          if (wordIndex !== -1) {
-            this.words[wordIndex] = {
-              ...this.words[wordIndex],
-              remark: remarkValue
-            };
-          }
-
-          // Update the DOM element for this word in the word list
-          const wordItem = document.querySelector(`[data-word-id="${word.id}"]`) as HTMLElement;
-          if (wordItem) {
-            // Find the existing remark display
-            const existingRemark = wordItem.querySelector('.text-orange-600');
-            if (existingRemark) {
-              if (remarkValue && remarkValue.trim()) {
-                // Update existing remark
-                const remarkText = existingRemark.querySelector('.truncate') as HTMLElement;
-                if (remarkText) {
-                  remarkText.textContent = remarkValue;
-                }
-              } else {
-                // Remove remark if it's empty
-                existingRemark.remove();
-              }
-            } else if (remarkValue && remarkValue.trim()) {
-              // Add new remark if it didn't exist before
-              const remarkHtml = `
-                <div class="text-xs text-orange-600 mt-1 italic flex items-center pr-1">
-                  <svg class="w-3 h-3 mr-1 flex-shrink-0 text-violet-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M11 5H6a2 2 0 0 0-2 2v11a2 2 0 0 0 2 2h11a2 2 0 0 0 2-2v-5m-1.414-9.414a2 2 0 1 1 2.828 2.828L11.828 15H9v-2.828z"/>
-                  </svg>
-                  <span class="truncate pr-0.5">${remarkValue}</span>
-                </div>
-              `;
-
-              // Find the description element and add remark after it
-              const description = wordItem.querySelector('.text-slate-500');
-              if (description && description.parentNode) {
-                const remarkDiv = document.createElement('div');
-                remarkDiv.innerHTML = remarkHtml;
-                description.parentNode.insertBefore(remarkDiv.firstElementChild!, description.nextSibling);
-              }
-            }
-          }
-
-          // Also update in associated words list if it exists there
-          const associatedWordItem = document.querySelector(`#associated-list [data-word-id="${word.id}"]`) as HTMLElement;
-          if (associatedWordItem) {
-            const existingRemark = associatedWordItem.querySelector('.text-orange-600');
-            if (existingRemark) {
-              if (remarkValue && remarkValue.trim()) {
-                // Update existing remark
-                const remarkText = existingRemark.querySelector('.truncate') as HTMLElement;
-                if (remarkText) {
-                  remarkText.textContent = remarkValue;
-                }
-              } else {
-                // Remove remark if it's empty
-                existingRemark.remove();
-              }
-            } else if (remarkValue && remarkValue.trim()) {
-              // Add new remark if it didn't exist before
-              const remarkHtml = `
-                <div class="text-xs text-orange-600 mt-1 italic flex items-center pr-1">
-                  <svg class="w-3 h-3 mr-1 flex-shrink-0 text-violet-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M11 5H6a2 2 0 0 0-2 2v11a2 2 0 0 0 2 2h11a2 2 0 0 0 2-2v-5m-1.414-9.414a2 2 0 1 1 2.828 2.828L11.828 15H9v-2.828z"/>
-                  </svg>
-                  <span class="truncate pr-0.5">${remarkValue}</span>
-                </div>
-              `;
-
-              // Find the description element and add remark after it
-              const description = associatedWordItem.querySelector('.text-slate-500');
-              if (description && description.parentNode) {
-                const remarkDiv = document.createElement('div');
-                remarkDiv.innerHTML = remarkHtml;
-                description.parentNode.insertBefore(remarkDiv.firstElementChild!, description.nextSibling);
-              }
-            }
-          }
-
-          // Re-render word details to show updated remark
-          await this.wordRenderer.renderWordDetails(this.currentWord);
-
-          if (remarkValue) {
-            this.toastManager.showSuccess('Remark saved successfully');
-          } else {
-            this.toastManager.showSuccess('Remark removed');
-          }
-        } else {
-          this.toastManager.showError('Failed to save remark');
-          // Restore original display
-          await this.wordRenderer.renderWordDetails(word);
-        }
-      } catch (error) {
-        console.error('Error saving remark:', error);
-        this.toastManager.showError('Failed to save remark');
-        // Restore original display
-        await this.wordRenderer.renderWordDetails(word);
-      }
-    };
-
-    const cancelEdit = () => {
-      if (saved) return;
-      // Restore original display
-      if (remarkDisplay) {
-        remarkContainer.replaceChild(remarkDisplay, input);
-      } else {
-        // If there was no original display, remove the input and the remark section
-        remarkContainer.remove();
-      }
-    };
-
-    // Handle key events
-    input.addEventListener('keydown', async (e) => {
-      if (e.key === 'Enter') {
-        e.preventDefault();
-        await saveRemark();
-      } else if (e.key === 'Escape') {
-        e.preventDefault();
-        cancelled = true;
-        cancelEdit();
-      }
-    });
-
-    // Handle blur (clicking outside)
-    input.addEventListener('blur', async () => {
-      // Small delay to allow for button clicks
-      setTimeout(async () => {
-        if (!saved && !cancelled) {
-          await saveRemark();
-        }
-      }, 150);
-    });
-  }
-
-  // Word Import methods
-  private selectedImportFile: File | null = null;
-
-  private showImportWordsModal(): void {
-    const modal = document.getElementById('import-words-modal')!;
-    if (modal) {
-      modal.classList.remove('hidden');
-    }
-  }
-
-  private hideImportWordsModal(): void {
-    const modal = document.getElementById('import-words-modal')!;
-    if (modal) {
-      modal.classList.add('hidden');
-    }
-    this.selectedImportFile = null;
-    this.updateImportUI();
-  }
-
-  private selectImportFile(): void {
-    const input = document.createElement('input');
-    input.type = 'file';
-    input.accept = '.txt,.md,.csv';
-    input.onchange = (e) => {
-      const target = e.target as HTMLInputElement;
-      const file = target.files?.[0];
-      if (file) {
-        this.selectedImportFile = file;
-        this.updateImportUI();
-      }
-    };
-    input.click();
-  }
-
-  private updateImportUI(): void {
-    const fileNameElement = document.getElementById('import-file-name') as HTMLElement;
-    const startBtn = document.getElementById('start-import-btn') as HTMLButtonElement;
-
-    if (fileNameElement && startBtn) {
-      if (this.selectedImportFile) {
-        fileNameElement.textContent = this.selectedImportFile.name;
-        startBtn.disabled = false;
-      } else {
-        fileNameElement.textContent = 'No file selected';
-        startBtn.disabled = true;
-      }
-    }
-  }
-
-  private async startWordImport(): Promise<void> {
-    if (!this.selectedImportFile || this.isImporting) {
-      return;
-    }
-
-    try {
-      const content = await this.readFileContent(this.selectedImportFile);
-      this.showImportProgressOverlay();
-
-      const callbacks: ImportCallbacks = {
-        onProgress: (progress: ImportProgress) => {
-          this.updateImportProgress(progress);
-        },
-        onComplete: (progress: ImportProgress) => {
-          this.handleImportComplete(progress);
-        },
-        onError: (progress: ImportProgress) => {
-          this.handleImportError(progress);
-        },
-        onCancel: (progress?: ImportProgress) => {
-          this.handleImportCancel(progress);
-        },
-      };
-
-      this.isImporting = true;
-      await this.wordImportService.startImport(content, callbacks);
-    } catch (error) {
-      console.error('Error starting import:', error);
-      this.toastManager.showError('Failed to start import');
-    }
-  }
-
-  private cancelWordImport(): void {
-    if (!this.isImporting) return;
-
-    this.wordImportService.cancelImport();
-  }
-
-  private async readFileContent(file: File): Promise<string> {
-    return new Promise((resolve, reject) => {
-      const reader = new FileReader();
-      reader.onload = (e) => {
-        const content = e.target?.result as string;
-        resolve(content);
-      };
-      reader.onerror = () => reject(new Error('Failed to read file'));
-      reader.readAsText(file);
-    });
-  }
-
-  private showImportProgressOverlay(): void {
-    const overlay = document.getElementById('import-progress-overlay')!;
-    if (overlay) {
-      overlay.classList.remove('hidden');
-    }
-  }
-
-  private hideImportProgressOverlay(): void {
-    const overlay = document.getElementById('import-progress-overlay')!;
-    if (overlay) {
-      overlay.classList.add('hidden');
-    }
-  }
-
-  private updateImportProgress(progress: ImportProgress): void {
-    const progressText = document.getElementById('import-progress-text')!;
-    const progressBar = document.getElementById('import-progress-bar') as HTMLDivElement;
-
-    if (progressText) {
-      progressText.textContent = `${progress.current}/${progress.total} - ${progress.currentWord || 'Processing...'}`;
-    }
-
-    if (progressBar) {
-      const percentage = progress.total > 0 ? (progress.current / progress.total) * 100 : 0;
-      progressBar.style.width = `${percentage}%`;
-    }
-  }
-
-  private handleImportComplete(progress: ImportProgress): void {
-    this.isImporting = false;
-    this.hideImportProgressOverlay();
-
-    // Show completion modal
-    const modal = document.getElementById('import-complete-modal')!;
-    const messageElement = document.getElementById('import-complete-message')!;
-    const okBtn = document.getElementById('import-complete-ok') as HTMLButtonElement;
-
-    if (messageElement) {
-      const successCount = progress.total - progress.errors.length - progress.skipped;
-      let html = `<span>Import completed!</span><br><span>${successCount}/${progress.total} words imported successfully.</span>`;
-
-      if (progress.skipped > 0) {
-        html += `<br><span>${progress.skipped} words were skipped (already exist).</span>`;
-      }
-
-      if (progress.errors.length > 0) {
-        html += `<br><span>${progress.errors.length} words had errors.</span>`;
-      }
-
-      messageElement.innerHTML = html;
-    }
-
-    if (modal) {
-      modal.classList.remove('hidden');
-    }
-
-    // Refresh word list
-    this.loadWords();
-
-    if (okBtn) {
-      okBtn.onclick = () => {
-        if (modal) {
-          modal.classList.add('hidden');
-        }
-      };
-    }
-  }
-
-  private handleImportError(progress: ImportProgress): void {
-    this.isImporting = false;
-    this.hideImportProgressOverlay();
-
-    // If we have progress info and at least one word was successfully imported, reload the word list
-    if (progress && progress.success > 0) {
-      console.log(`Import failed but ${progress.success} words were successfully imported. Reloading word list...`);
-      this.loadWords();
-    }
-
-    // Show detailed completion modal instead of simple toast
-    this.showImportErrorModal(progress);
-  }
-
-  private handleImportCancel(progress?: ImportProgress): void {
-    this.isImporting = false;
-    this.hideImportProgressOverlay();
-
-    // If we have progress info and at least one word was successfully imported, reload the word list
-    if (progress && progress.success > 0) {
-      console.log(`Import cancelled but ${progress.success} words were successfully imported. Reloading word list...`);
-      this.loadWords();
-    }
-
-    this.toastManager.showWarning('Import cancelled');
-  }
-
-  private showImportErrorModal(progress: ImportProgress): void {
-    const modal = document.getElementById('import-complete-modal')!;
-    const messageElement = document.getElementById('import-complete-message')!;
-    const iconContainer = modal.querySelector('.w-16.h-16') as HTMLElement;
-    const titleElement = modal.querySelector('h3') as HTMLElement;
-    const okBtn = document.getElementById('import-complete-ok') as HTMLButtonElement;
-
-    if (messageElement && iconContainer && titleElement) {
-      // Change icon to warning for partial success
-      iconContainer.innerHTML = `
-        <svg class="w-8 h-8 text-amber-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-          <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-2.5L13.732 4c-.77-.833-1.964-.833-2.732 0L3.732 16.5c-.77.833.192 2.5 1.732 2.5z"></path>
-        </svg>
-      `;
-
-      // Update title for partial success
-      titleElement.textContent = 'Import Stopped - Partial Success';
-
-      // Calculate statistics
-      const successfulCount = progress.success || 0;
-      const failedCount = progress.errors?.length || 0;
-      const skippedCount = progress.skipped || 0;
-      const remainingCount = (progress.total || 0) - (progress.current || 0);
-      const failedWord = progress.currentWord || 'unknown word';
-
-      // Build detailed message
-      let html = `<div class="text-left space-y-2">`;
-
-      if (successfulCount > 0) {
-        html += `<div class="flex items-center text-green-600">
-          <svg class="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M5 13l4 4L19 7"></path>
-          </svg>
-          <span class="font-medium">${successfulCount} words successfully imported</span>
-        </div>`;
-      }
-
-      if (skippedCount > 0) {
-        html += `<div class="flex items-center text-blue-600">
-          <svg class="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M13 7h8m0 0v8m0-8l-8 8-4-4-6 6"></path>
-          </svg>
-          <span class="font-medium">${skippedCount} words were skipped (already exist)</span>
-        </div>`;
-      }
-
-      if (failedCount > 0) {
-        const error = progress.errors?.[0] || 'Unknown error';
-        html += `<div class="flex items-center text-red-600">
-          <svg class="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"></path>
-          </svg>
-          <span class="font-medium">Failed on "${failedWord}": ${error}</span>
-        </div>`;
-      }
-
-      if (remainingCount > 0) {
-        html += `<div class="flex items-center text-amber-600">
-          <svg class="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"></path>
-          </svg>
-          <span class="font-medium">${remainingCount} words remaining unprocessed</span>
-        </div>`;
-      }
-
-      html += `</div>`;
-      messageElement.innerHTML = html;
-    }
-
-    if (modal) {
-      modal.classList.remove('hidden');
-    }
-
-    if (okBtn) {
-      okBtn.onclick = () => {
-        if (modal) {
-          modal.classList.add('hidden');
-        }
-      };
-    }
+    this.modalManager.hideHowtoModal();
   }
 
   private toggleApiKeyVisibility(): void {
-    const apiKeyInput = document.getElementById('api-key') as HTMLInputElement;
-    const eyeIcon = document.getElementById('eye-icon') as HTMLElement;
-    const eyeOffIcon = document.getElementById('eye-off-icon') as HTMLElement;
-
-    if (apiKeyInput && eyeIcon && eyeOffIcon) {
-      if (apiKeyInput.type === 'password') {
-        // Show password
-        apiKeyInput.type = 'text';
-        eyeIcon.classList.add('hidden');
-        eyeOffIcon.classList.remove('hidden');
-      } else {
-        // Hide password
-        apiKeyInput.type = 'password';
-        eyeIcon.classList.remove('hidden');
-        eyeOffIcon.classList.add('hidden');
-      }
-    }
+    this.modalManager.toggleApiKeyVisibility();
   }
 
   private async handleSortToggle(): Promise<void> {
-    // Toggle sort order
-    this.sortOrder = this.sortOrder === 'desc' ? 'asc' : 'desc';
-
-    // Save to StoreManager
-    await this.saveSortOrder();
-
-    // Update sort button icon
-    this.updateSortButtonIcon();
-
-    // Reload words with new sort order
-    await this.loadWords();
+    await this.wordManager.handleSortToggle();
   }
 
-  private async saveSortOrder(): Promise<void> {
-    try {
-      // Save to electron-store via IPC
-      await window.electronAPI.saveSortOrder(this.sortOrder);
-    } catch (error) {
-      console.error('Error saving sort order:', error);
-    }
-  }
 
-  private async loadSortOrder(): Promise<void> {
-    try {
-      // Load from electron-store via IPC
-      const savedSortOrder = await window.electronAPI.loadSortOrder();
-      if (savedSortOrder) {
-        this.sortOrder = savedSortOrder;
-      }
-    } catch (error) {
-      console.error('Error loading sort order:', error);
-      // Keep default 'desc'
-    }
-  }
-
-  private updateSortButtonIcon(): void {
-    const sortBtn = document.getElementById('sort-btn') as HTMLButtonElement;
-    if (!sortBtn) return;
-
-    const iconContainer = sortBtn.querySelector('.sort-icon');
-    if (!iconContainer) return;
-
-    if (this.sortOrder === 'desc') {
-      // Down arrow for descending (newest first) - more elegant design
-      iconContainer.innerHTML = `
-        <svg class="w-4 h-4" viewBox="0 0 6.4 6.4" xmlns="http://www.w3.org/2000/svg"><path d="m5.741 4.341-1 1-.001.001-.014.013-.007.005-.009.006-.009.005-.008.005-.009.004-.009.004-.009.003-.019.005-.01.002-.01.002-.009.001h-.039l-.009-.001-.011-.002-.01-.002-.009-.002-.01-.003-.009-.003-.009-.004-.009-.004-.008-.005-.009-.005-.008-.006-.008-.006-.013-.011-.002-.002-1-1a.2.2 0 0 1 .283-.283l.662.659V2.8a.2.2 0 0 1 .4 0v1.917l.659-.658a.2.2 0 1 1 .283.283M3 3H1.2a.2.2 0 0 0 0 .4H3A.2.2 0 1 0 3 3M1.2 1.8h3.4a.2.2 0 0 0 0-.4H1.2a.2.2 0 1 0 0 .4m1.4 2.8H1.2a.2.2 0 0 0 0 .4h1.4a.2.2 0 0 0 0-.4"/></svg>
-      `;
-      sortBtn.title = 'Newest first → Oldest first';
-    } else {
-      // Up arrow for ascending (oldest first) - more elegant design
-      iconContainer.innerHTML = `
-        <svg class="w-4 h-4" viewBox="0 0 6.4 6.4" xmlns="http://www.w3.org/2000/svg"><path d="M5.741 2.341a.2.2 0 0 1-.283 0L4.8 1.683V3.6a.2.2 0 0 1-.4 0V1.683l-.659.658a.2.2 0 0 1-.283-.283l1-1 .002-.002.013-.011.007-.006.008-.006.009-.005.008-.005.009-.004.009-.004.009-.003.01-.004.008-.002.011-.003.009-.001.10-.001L4.595 1h.009l.015.001.01.002.009.001.011.003.008.002.01.004.008.003.009.004.009.004.009.005.008.005.009.007.007.005.014.013.001.001 1 1a.2.2 0 0 1 0 .283M1.2 3.4H3A.2.2 0 1 0 3 3H1.2a.2.2 0 0 0 0 .4m0-1.6h1.4a.2.2 0 0 0 0-.4H1.2a.2.2 0 1 0 0 .4m3.4 2.8H1.2a.2.2 0 0 0 0 .4h3.4a.2.2 0 0 0 0-.4"/></svg>
-      `;
-      sortBtn.title = 'Oldest first → Newest first';
-    }
-  }
-
-  // Auto-generate word if it doesn't exist
-  private async autoGenerateWord(wordName: string): Promise<void> {
-    try {
-      // Check if we're already generating something
-      if (this.isGenerating) {
-        this.toastManager.showWarning('Please wait for current generation to complete');
-        return;
-      }
-
-      // Set the word in input field
-      const wordInput = document.getElementById('word-input') as HTMLInputElement;
-      wordInput.value = wordName;
-
-      // Show loading message
-      this.toastManager.showInfo(`Generating word: ${wordName}...`);
-
-      // Trigger generation using existing logic
-      await this.handleGenerate();
-
-      // The word should now be generated and selected
-      this.toastManager.showSuccess(`Word "${wordName}" generated and selected!`);
-
-    } catch (error) {
-      console.error('Error auto-generating word:', error);
-      this.toastManager.showError(`Failed to generate word: ${wordName}`);
-    }
-  }
 
   // Protocol handlers
   private async handleProtocolNavigateWord(wordName: string): Promise<void> {
-    try {
-      console.log('🎯 Renderer: Handling protocol navigation to word:', wordName);
-
-      // Try to find the word by name
-      const word = await this.wordService.getWordByName(wordName);
-      if (word) {
-        // Word found, select it
-        this.selectWord(word);
-        this.toastManager.showSuccess(`Navigated to word: ${wordName}`);
-      } else {
-        // Word not found, auto-generate it!
-        console.log('🎯 Renderer: Word not found, auto-generating:', wordName);
-        await this.autoGenerateWord(wordName);
-      }
-    } catch (error) {
-      console.error('Error handling protocol navigation:', error);
-      this.toastManager.showError('Failed to navigate to word');
-    }
+    await this.protocolManager.handleProtocolNavigateWord(wordName);
   }
 
   private async handleProtocolSwitchProfile(profileName: string): Promise<void> {
-    try {
-      console.log('Handling protocol profile switch to:', profileName);
-
-      // Check if the profile exists
-      const profiles = await this.profileService.getProfiles();
-      if (profiles.includes(profileName)) {
-        // Show loading overlay
-        this.showLoadingOverlay();
-
-        try {
-          const success = await this.profileService.switchProfile(profileName);
-          if (success) {
-            // Successful switch - load words normally
-            this.resetUIForProfileSwitch();
-            await this.loadWords();
-            this.toastManager.showSuccess(`Switched to profile: ${profileName}`);
-          } else {
-            this.toastManager.showError(`Failed to switch to profile: ${profileName}`);
-          }
-        } catch (error) {
-          console.error('Error switching profile:', error);
-          this.toastManager.showError(`Failed to switch to profile: ${profileName}`);
-        } finally {
-          setTimeout(() => {
-            this.hideLoadingOverlay();
-          }, 500);
-        }
-      } else {
-        this.toastManager.showError(`Profile "${profileName}" not found`);
-      }
-    } catch (error) {
-      console.error('Error handling protocol profile switch:', error);
-      this.toastManager.showError('Failed to switch profile');
-    }
-  }
-
-  // Google Drive modal methods
-  private showExportChoiceModal(): void {
-    const modal = document.getElementById('export-choice-modal')!;
-    if (modal) {
-      modal.classList.remove('hidden');
-    }
-  }
-
-  private hideExportChoiceModal(): void {
-    const modal = document.getElementById('export-choice-modal')!;
-    if (modal) {
-      modal.classList.add('hidden');
-    }
-  }
-
-  private showImportChoiceModal(): void {
-    const modal = document.getElementById('import-choice-modal')!;
-    if (modal) {
-      modal.classList.remove('hidden');
-    }
-  }
-
-  private hideImportChoiceModal(): void {
-    const modal = document.getElementById('import-choice-modal')!;
-    if (modal) {
-      modal.classList.add('hidden');
-    }
-  }
-
-  private async showGoogleDriveFilePicker(): Promise<void> {
-    try {
-      // Reset selection state when opening modal
-      this.selectedGoogleDriveFile = null;
-      this.updateGoogleDriveImportUI();
-
-      // Check if we're authenticated first
-      const authStatus = await window.electronAPI.googleIsAuthenticated();
-      if (!authStatus.authenticated) {
-        // Trigger authentication
-        const authResult = await window.electronAPI.googleAuthenticate();
-        if (!authResult.success) {
-          this.toastManager.showError('Authentication required for Google Drive access');
-          return;
-        }
-      }
-
-      // Show loading in the file picker
-      const filesList = document.getElementById('google-drive-files-list')!;
-      if (filesList) {
-        filesList.innerHTML = `
-          <div class="text-center text-slate-500 py-8">
-            <div class="animate-spin rounded-full h-8 w-8 border-b-2 border-blue-500 mx-auto mb-4"></div>
-            <p>Loading files...</p>
-          </div>
-        `;
-      }
-
-      // Show the modal
-      const modal = document.getElementById('google-drive-picker-modal')!;
-      if (modal) {
-        modal.classList.remove('hidden');
-      }
-
-      // Load files from Google Drive
-      const result = await window.electronAPI.googleDriveListFiles();
-      if (result.success && result.files) {
-        this.renderGoogleDriveFiles(result.files);
-      } else {
-        this.showGoogleDriveError('Failed to load Google Drive files');
-      }
-    } catch (error) {
-      console.error('Failed to show Google Drive file picker:', error);
-      this.showGoogleDriveError('Failed to load Google Drive files');
-    }
-  }
-
-  private hideGoogleDriveFilePicker(): void {
-    const modal = document.getElementById('google-drive-picker-modal')!;
-    if (modal) {
-      modal.classList.add('hidden');
-    }
-  }
-
-  private hideGoogleDriveUploadModal(): void {
-    const modal = document.getElementById('google-drive-upload-modal')!;
-    if (modal) {
-      modal.classList.add('hidden');
-    }
-  }
-
-  private renderGoogleDriveFiles(files: any[]): void {
-    const filesList = document.getElementById('google-drive-files-list')!;
-    if (!filesList) return;
-
-    if (files.length === 0) {
-      filesList.innerHTML = `
-        <div class="text-center text-slate-500 py-8">
-          <svg class="w-12 h-12 mx-auto mb-4 text-slate-300" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"></path>
-          </svg>
-          <p>No profiles found in Google Drive</p>
-          <p class="text-sm mt-2">Upload a profile first to see it here</p>
-        </div>
-      `;
-      return;
-    }
-
-    // Clear the list first
-    filesList.innerHTML = '';
-
-    // Create file elements individually to avoid HTML injection issues
-    files.forEach(file => {
-      const isSelected = this.selectedGoogleDriveFile && this.selectedGoogleDriveFile.id === file.id;
-
-      // Create the main container
-      const fileContainer = document.createElement('div');
-      fileContainer.className = 'relative group';
-
-      // Create the file button
-      const fileButton = document.createElement('button');
-      fileButton.className = `w-full p-4 ${isSelected ? 'bg-blue-50 border-blue-300' : 'bg-white/80 hover:bg-white/90'} border ${isSelected ? 'border-blue-300' : 'border-slate-200'} rounded-lg transition-all duration-200 hover:shadow-md google-drive-file-btn`;
-      fileButton.setAttribute('data-file-id', file.id);
-
-      // Create the content structure
-      fileButton.innerHTML = `
-        <div class="flex items-center">
-          <svg class="w-8 h-8 mr-3 text-blue-500" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-            <path d="M12 2l3.09 6.26L22 9.27l-5 4.87 1.18 6.88L12 17.77l-6.18 3.25L7 14.14 2 9.27l6.91-1.01L12 2z" fill="#4285F4"/>
-          </svg>
-          <div class="text-left flex-1">
-            <h4 class="font-medium ${isSelected ? 'text-blue-800' : 'text-slate-800'} truncate"></h4>
-            <p class="text-sm text-slate-500">Modified: ${new Date(file.modifiedTime).toLocaleDateString()}</p>
-            ${file.size ? `<p class="text-sm text-slate-500">Size: ${this.formatFileSize(parseInt(file.size))}</p>` : ''}
-          </div>
-        </div>
-      `;
-
-      // Set the file name using textContent to avoid HTML injection
-      const fileNameElement = fileButton.querySelector('h4');
-      if (fileNameElement) {
-        fileNameElement.textContent = file.name;
-      }
-
-      // Create the delete button
-      const deleteButton = document.createElement('button');
-      deleteButton.className = 'absolute top-1/2 right-4 -translate-y-1/2 p-1.5 bg-red-500 hover:bg-red-600 text-white rounded-full opacity-0 group-hover:opacity-100 transition-opacity duration-200 google-drive-delete-btn';
-      deleteButton.setAttribute('data-file-id', file.id);
-      deleteButton.setAttribute('data-file-name', file.name);
-      deleteButton.title = 'Delete file';
-      deleteButton.innerHTML = `
-        <svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-          <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16"></path>
-        </svg>
-      `;
-
-      // Add click handler for file button
-      fileButton.addEventListener('click', async (e) => {
-        const target = e.currentTarget as HTMLElement;
-        const fileId = target.getAttribute('data-file-id');
-        if (fileId) {
-          // Find the file object
-          const selectedFile = files.find(f => f.id === fileId);
-          if (selectedFile) {
-            this.selectedGoogleDriveFile = selectedFile;
-            this.updateGoogleDriveImportUI();
-            // Re-render to show selection
-            this.renderGoogleDriveFiles(files);
-          }
-        }
-      });
-
-      // Add click handler for delete button
-      deleteButton.addEventListener('click', async (e) => {
-        e.stopPropagation(); // Prevent triggering the file selection
-        const fileId = deleteButton.getAttribute('data-file-id');
-        const fileName = deleteButton.getAttribute('data-file-name');
-        if (fileId && fileName) {
-          await this.handleGoogleDriveFileDelete(fileId, fileName, files);
-        }
-      });
-
-      // Assemble the elements
-      fileContainer.appendChild(fileButton);
-      fileContainer.appendChild(deleteButton);
-      filesList.appendChild(fileContainer);
-    });
-  }
-
-  private updateGoogleDriveImportUI(): void {
-    const selectedFileInfo = document.getElementById('selected-file-info') as HTMLElement;
-    const selectedFileName = document.getElementById('selected-file-name') as HTMLElement;
-    const importBtn = document.getElementById('import-google-drive-file') as HTMLButtonElement;
-
-    if (this.selectedGoogleDriveFile) {
-      if (selectedFileInfo && selectedFileName) {
-        selectedFileInfo.classList.remove('hidden');
-        selectedFileName.textContent = this.selectedGoogleDriveFile.name;
-      }
-      if (importBtn) {
-        importBtn.disabled = false;
-        importBtn.classList.remove('disabled:bg-slate-400', 'disabled:cursor-not-allowed');
-        importBtn.classList.add('bg-blue-500', 'hover:bg-blue-600');
-      }
-    } else {
-      if (selectedFileInfo) {
-        selectedFileInfo.classList.add('hidden');
-      }
-      if (importBtn) {
-        importBtn.disabled = true;
-        importBtn.classList.remove('bg-blue-500', 'hover:bg-blue-600');
-        importBtn.classList.add('disabled:bg-slate-400', 'disabled:cursor-not-allowed');
-      }
-    }
-  }
-
-  private async performGoogleDriveImport(fileId: string): Promise<void> {
-    try {
-      this.showLoadingOverlay();
-      this.hideGoogleDriveFilePicker();
-
-      const result = await window.electronAPI.googleDriveDownloadDatabase(fileId);
-
-      if (result.success) {
-        this.toastManager.showSuccess(result.message);
-
-        // Refresh profiles and switch to the new one
-        await this.profileService.loadProfiles();
-        if (result.profileName) {
-          this.profileService.setCurrentProfile(result.profileName);
-          const profileSelect = document.getElementById('profile-select') as HTMLSelectElement;
-          profileSelect.value = result.profileName;
-
-          this.resetUIForProfileSwitch();
-          await this.loadWords();
-        }
-      } else {
-        this.toastManager.showError(result.message);
-      }
-    } catch (error) {
-      console.error('Failed to download Google Drive file:', error);
-      this.toastManager.showError('Failed to download file from Google Drive');
-    } finally {
-      this.hideLoadingOverlay();
-      // Reset selection
-      this.selectedGoogleDriveFile = null;
-    }
-  }
-
-  private showGoogleDriveError(message: string): void {
-    const filesList = document.getElementById('google-drive-files-list')!;
-    if (filesList) {
-      filesList.innerHTML = `
-        <div class="text-center text-red-500 py-8">
-          <svg class="w-12 h-12 mx-auto mb-4 text-red-300" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"></path>
-          </svg>
-          <p class="font-medium">Error</p>
-          <p class="text-sm mt-2">${message}</p>
-        </div>
-      `;
-    }
+    await this.protocolManager.handleProtocolSwitchProfile(profileName);
   }
 
 
-
-  private async handleExportToGoogleDrive(): Promise<void> {
-    try {
-      this.showLoadingOverlay();
-
-      // Check if we're authenticated first
-      const authStatus = await window.electronAPI.googleIsAuthenticated();
-      if (!authStatus.authenticated) {
-        // Trigger authentication
-        const authResult = await window.electronAPI.googleAuthenticate();
-        if (!authResult.success) {
-          this.toastManager.showError('Authentication required for Google Drive access');
-          return;
-        }
-      }
-
-      const result = await window.electronAPI.googleDriveUploadDatabase();
-
-      if (result.success) {
-        // Show the upload success modal with ALL uploaded files
-        // This gives users a complete view of their backup history
-        await this.showGoogleDriveUploadSuccess(result.fileId);
-      } else {
-        this.toastManager.showError(result.message);
-      }
-    } catch (error) {
-      console.error('Failed to export to Google Drive:', error);
-      this.toastManager.showError('Failed to export to Google Drive');
-    } finally {
-      this.hideLoadingOverlay();
-    }
-  }
-
-  private async showGoogleDriveUploadSuccess(justUploadedFileId?: string): Promise<void> {
-    const filesList = document.getElementById('google-drive-uploaded-files-list')!;
-    if (!filesList) return;
-
-    try {
-      // Show loading state
-      filesList.innerHTML = `
-        <div class="text-center text-slate-500 py-8">
-          <div class="animate-spin rounded-full h-8 w-8 border-b-2 border-blue-500 mx-auto mb-4"></div>
-          <p>Loading your backup files...</p>
-        </div>
-      `;
-
-      // Show the modal
-      const modal = document.getElementById('google-drive-upload-modal')!;
-      if (modal) {
-        modal.classList.remove('hidden');
-      }
-
-      // Load all EverEtch files from Google Drive
-      const result = await window.electronAPI.googleDriveListFiles();
-      if (result.success && result.files) {
-        this.renderGoogleDriveUploadedFiles(result.files, justUploadedFileId);
-      } else {
-        filesList.innerHTML = `
-          <div class="text-center text-slate-500 py-4">
-            <p>Unable to load backup files</p>
-          </div>
-        `;
-      }
-    } catch (error) {
-      console.error('Failed to load uploaded files:', error);
-      filesList.innerHTML = `
-        <div class="text-center text-slate-500 py-4">
-          <p>Unable to load backup files</p>
-        </div>
-      `;
-    }
-  }
-
-  private renderGoogleDriveUploadedFiles(files: any[], justUploadedFileId?: string): void {
-    const filesList = document.getElementById('google-drive-uploaded-files-list')!;
-    if (!filesList) return;
-
-    if (files.length === 0) {
-      filesList.innerHTML = `
-        <div class="text-center text-slate-500 py-4">
-          <p>No backup files found in Google Drive</p>
-        </div>
-      `;
-      return;
-    }
-
-    // Clear the list first
-    filesList.innerHTML = '';
-
-    // Sort files by modified time (newest first)
-    const sortedFiles = files.sort((a, b) =>
-      new Date(b.modifiedTime).getTime() - new Date(a.modifiedTime).getTime()
-    );
-
-    // Create file elements individually to avoid HTML injection issues
-    sortedFiles.forEach(file => {
-      const isJustUploaded = file.id === justUploadedFileId;
-
-      // Create the main container
-      const fileContainer = document.createElement('div');
-      fileContainer.className = `flex items-center p-3 ${isJustUploaded ? 'bg-green-50 border border-green-200' : 'bg-slate-50'} rounded-lg`;
-
-      // Create the content structure
-      fileContainer.innerHTML = `
-        <svg class="w-6 h-6 mr-3 text-blue-500" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-          <path d="M12 2l3.09 6.26L22 9.27l-5 4.87 1.18 6.88L12 17.77l-6.18 3.25L7 14.14 2 9.27l6.91-1.01L12 2z" fill="#4285F4"/>
-        </svg>
-        <div class="text-left flex-1">
-          <div class="flex items-center">
-            <h4 class="font-medium text-slate-800 truncate"></h4>
-            ${isJustUploaded ? '<span class="ml-2 px-2 py-1 text-xs bg-green-100 text-green-700 rounded-full font-medium">Just uploaded</span>' : ''}
-          </div>
-          <p class="text-sm text-slate-500">Modified: ${new Date(file.modifiedTime).toLocaleDateString()}</p>
-          ${file.size ? `<p class="text-sm text-slate-500">Size: ${this.formatFileSize(parseInt(file.size))}</p>` : ''}
-        </div>
-        <svg class="w-5 h-5 ${isJustUploaded ? 'text-green-500' : 'text-slate-400'}" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-          <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M5 13l4 4L19 7"></path>
-        </svg>
-      `;
-
-      // Set the file name using textContent to avoid HTML injection
-      const fileNameElement = fileContainer.querySelector('h4');
-      if (fileNameElement) {
-        fileNameElement.textContent = file.name;
-      }
-
-      filesList.appendChild(fileContainer);
-    });
-  }
-
-  private formatFileSize(bytes: number): string {
-    const units = ['B', 'KB', 'MB', 'GB'];
-    let size = bytes;
-    let unitIndex = 0;
-
-    while (size >= 1024 && unitIndex < units.length - 1) {
-      size /= 1024;
-      unitIndex++;
-    }
-
-    return `${size.toFixed(1)} ${units[unitIndex]}`;
-  }
-
-  private async handleGoogleDriveFileDelete(fileId: string, fileName: string, files: any[]): Promise<void> {
-    const confirmed = confirm(`Are you sure you want to delete "${fileName}" from Google Drive?\n\nThis action cannot be undone.`);
-
-    if (!confirmed) {
-      return;
-    }
-
-    try {
-      // Show loading overlay for delete operation
-      this.showLoadingOverlay();
-
-      const result = await window.electronAPI.googleDriveDeleteFile(fileId);
-
-      if (result.success) {
-        this.toastManager.showSuccess(`File "${fileName}" deleted successfully`);
-
-        // Remove the file from the current list and refresh the UI
-        const updatedFiles = files.filter(file => file.id !== fileId);
-        this.renderGoogleDriveFiles(updatedFiles);
-
-        // If the deleted file was selected, clear the selection
-        if (this.selectedGoogleDriveFile && this.selectedGoogleDriveFile.id === fileId) {
-          this.selectedGoogleDriveFile = null;
-          this.updateGoogleDriveImportUI();
-        }
-      } else {
-        this.toastManager.showError(result.message || 'Failed to delete file');
-      }
-    } catch (error) {
-      console.error('Failed to delete Google Drive file:', error);
-      this.toastManager.showError('Failed to delete file from Google Drive');
-    } finally {
-      // Hide loading overlay
-      setTimeout(() => {
-        this.hideLoadingOverlay();
-      }, 500);
-    }
-  }
 }
